This does not work.
Well I failed with the sort method. But after looking through the
"Possible Solutions" section, I can see that I was on the right track. I
am forced to move on for the sake of schedule. I chose to leave my
failed attempt here rather than copy and paste the possible solution so
that it can be seen that I'm on the right path.